+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ## [0.4.2] - 2025-04-09
 
+### Added
+
+- Add vendor compliance reports UI
+
 ### Changed
 
 - Simplified policy data model by removing version field and optimistic concurrency